refactor: move lifecycle types from cella-docker to cella-backend
Move LifecycleContext, ParsedLifecycle, parse_lifecycle_command, and
run_lifecycle_phase to cella-backend so the orchestrator can use them
without depending on cella-docker. The client field in LifecycleContext
changes from &DockerClient to &dyn ContainerBackend.

cella-docker re-exports these types for backward compatibility until
all consumers are migrated.

Ref: #15
