+++ crates/cella-backend/Cargo.toml
@@ -6,9 +6,12 @@ license.workspace = true
 
 [dependencies]
 chrono.workspace = true
+futures-util.workspace = true
 hex.workspace = true
+serde_json.workspace = true
 sha2.workspace = true
 thiserror.workspace = true
+tracing.workspace = true
 
 [lints]
 workspace = true
