+++ crates/cella-backend/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod error;
+pub mod lifecycle;
 pub mod names;
 pub mod traits;
 pub mod types;
@@ -7,6 +8,10 @@ pub use error::BackendError;
 pub use names::{
     compose_labels, compose_project_name, container_labels, container_name, image_name,
     image_name_with_features, worktree_labels,
+};
+pub use lifecycle::{
+    LifecycleContext, OutputCallback, ParsedLifecycle, parse_lifecycle_command,
+    run_lifecycle_phase,
 };
 pub use traits::{BoxFuture, ComposeBackend, ContainerBackend, Platform};
 pub use types::{
